requirements.txt
click>=8.0.0
python-dotenv>=0.19.0 